requirements.txt
numpy
onnxruntime
onnxruntime-genai-directml
huggingface-hub
soundfile
torch
silero-vad
sounddevice
simpleaudio
grpcio
zmq